Create Postfix to Infix Conversion.cpp
In this tutorial we will convert in Postfix Expression to a infix Expression using Stack Data structure. We will understand the Rules to convert an postfix expression to infix expression and also understand the pseudocode. Lastly we will write a C++ program to perform postfix to infix expression conversion.

Rules for Postfix to Infix using stack DS –
1.Scan POSTFIX expression from LEFT to RIGHT
2.IF the incoming symbol is a OPERAND, PUSH it onto the Stack
3.IF the incoming symbol is a OPERATOR, POP 2 OPERANDs from the Stack, ADD this incoming OPERATOR in between the 2 OPERANDs, ADD ‘(‘ & ‘)’ to the whole expression & PUSH this whole new expression string back into the Stack.
4.At the end POP and PRINT the full INFIX expression from the Stack.