Create Postfix to Prefix Conversion,cpp
In this tutorial we will convert in Postfix Expression to a Prefix Expression using Stack Data structure. We will understand the Rules to convert a Postfix expression to Prefix expression and also understand the pseudocode. Lastly we will write a C++ program to perform Postfix  to Prefix expression conversion.

Rules for Postfix to Prefix using stack DS –
1.Scan POSTFIX expression from LEFT to RIGHT
2.IF the incoming symbol is a OPERAND, PUSH it onto the Stack
3.IF the incoming symbol is a OPERATOR, POP 2 OPERANDs from the Stack, ADD this incoming OPERATOR in before the 2 OPERANDs & PUSH this whole new expression string back into the Stack.
4.At the end POP and PRINT the full PREFIX expression from the Stack.