Create Prefix to Postfix Conversion.cpp
In this tutorial we will convert in Prefix Expression to a Postfix Expression using Stack Data structure. We will understand the Rules to convert a Prefix expression to Postfix expression and also understand the pseudocode. Lastly we will write a C++ program to perform Prefix to Postfix expression conversion.

Rules for Prefix to Postfix using stack DS –
1.Scan PREFIX expression from RIGHT to LEFT
2.IF the incoming symbol is a OPERAND, PUSH it onto the Stack
3.IF the incoming symbol is a OPERATOR, POP 2 OPERANDs from the Stack, ADD this incoming OPERATOR in before the 2 OPERANDs & PUSH this whole new expression string back into the Stack.
4.At the end POP and PRINT the full POSTFIX expression from the Stack.